BUG: Fix 64-to-32 bit truncations, including a real overflow in fslio
31 -Wshorten-64-to-32 warnings.  Most were benign, one was not.

The real defect is FslSeekVolume():

    int offset;
    offset = fslio->niftiptr->iname_offset
             + vols * FslGetVolSize(fslio) * fslio->niftiptr->nbyper;
    return znzseek(fslio->fileptr, offset, SEEK_SET);

offset is an int, but it holds a byte position into the image file.  Seek
to any volume past 2GB and the multiplication overflows, so znzseek --
which takes a 64-bit znz_off_t -- receives a wrong and quite possibly
negative position.  offset is a local, so widening it to znz_off_t and
casting the operands changes no interface.  FslReadVolumes' volbytes had
the same shape and is now size_t.

The rest fall into two groups.

Internal helpers were widened to carry the value they were already being
handed.  In nifti2_io.c, nifti_read_extensions() and rci_alloc_mem() now
return int64_t rather than truncating their own int64_t results, and
nifti_read_next_extension() and nifti_check_extension() take an int64_t
`remain` rather than accepting an implicit narrowing from their caller.
nt_read_bricks() in nifti_tool.h likewise takes an int64_t length.  All
of these are static or live in an uninstalled header.

Where the narrowing target is an installed prototype it cannot be fixed,
only made explicit, so each such cast carries a comment saying why the
value cannot exceed the range and what would have to change otherwise:
nifti_image_load_bricks() returning a brick count as int,
nifti_read_subregion_image() returning a byte count as int,
nifti_read_ascii_image() taking an int header length, axio_num_tokens()
returning int, and FslSeekVolume()'s int return.  The remainder are
strlen() results assigned to int and fread()/znzread() counts, all of
them bounded by buffers a few hundred bytes long.

Verified with both gcc and clang, and nm -D, the installed declarations
and every installed macro are unchanged.

+++ cifti/afni_xml.c
@@ -282,7 +282,7 @@ axml_read_file(const char * fname, int read_data)
     if (reset_xml_buf(xd, &buf, &bsize))
       break;
 
-    blen = fread(buf, 1, bsize, fp);
+    blen = (unsigned)fread(buf, 1, bsize, fp);
 
     /* check for early termination */
     bshort = loc_strnlen(buf, blen);
@@ -378,7 +378,7 @@ axml_read_buf(const char * buf_in, int64_t bin_len)
     if (bin_remain >= bsize)
       blen = bsize;
     else
-      blen = bin_remain;
+      blen = (unsigned)bin_remain;
 
     if (blen > 0 && blen <= (unsigned)bsize)
     {
@@ -1099,7 +1099,7 @@ strip_whitespace(const char * str, int slen)
   if (!str || slen > 1024)
     return str;
 
-  len = strlen(str);
+  len = (int)strlen(str);
   if (slen > 0 && slen < len)
     len = slen;
   if (len <= 0)
@@ -1213,7 +1213,7 @@ disp_gen_text(afni_xml_control * xd, const char * header, const char * cdata, in
   if (len == wlen)
   {
     str = "whitespace"; /* just note the whitespace */
-    len = strlen(str);
+    len = (int)strlen(str);
   }
 
   show_depth(xd, 1);
